Added a test on the ScheduledService and adjusted Team Entity

+++ code/battle-manager/src/main/java/ckb/BattleManager/dto/output/BattleFinishedMessage.java
@@ -10,5 +10,5 @@ import java.util.List;
 @AllArgsConstructor
 public class BattleFinishedMessage {
     private Long idTournament;
-    private List<Pair<Long, Long>> pairsIdUserPoints;
+    private List<Pair<Long, Integer>> pairsIdUserPoints;
 }
